% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bank}
\alias{bank}
\title{Random arrival and service times.}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{X}{numbering of the customers}
\item{arrival_time}{arrival time, in seconds}
\item{service_time}{service duration, in seconds}
...
}
}
\usage{
bank
}
\description{
A data set containing the arrival times and service duration of 100 customers.
}
\keyword{datasets}
