% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{Multiserver}
\alias{Multiserver}
\title{Queuing simulation}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Input the whole column of the customers' arrival times.}

\item{ServiceTimes}{Input the whole column of the required service duration.}

\item{NumServers}{Input the number of total servers.}
}
\value{
A tibble of when each customers arrived, serving started and ended and which server was chosen.
}
\description{
This function allows you to put to test a first come, first serve queuing system. You decide how many people, how much service time they take and how many servers are available.
}
\examples{
Multiserver(Arrivals = bank$arrival_time, ServiceTimes = bank$service_time, NumServers = 2)

}
